+++ viewjeepdetails.php
@@ -102,7 +102,7 @@
                         <td>
                            <a href="moredetailsjeep.php?id=<?php echo $row["id"]; ?>" target="_blank" class="read">Read More</a>
                            <a href="updatejeepdetails.php?id=<?php echo $row["id"]; ?>" target="_blank" class="update">Update</a>
-                           <a href="deletejeep.php?id=<?php echo $row["id"]; ?>" target="_blank" class="delete">Delete</a>
+                           <a href="deletejeep.php?id=<?php echo $row["id"]; ?>" target="_blank" class="delete" onclick="return confirmDelete();">Delete</a>
                         </td>
                      </tr>
                   <?php
@@ -211,5 +211,14 @@
 
     </footer>
     <script src="./raminduform.js"></script>
+
+    <script>
+          function confirmDelete() {
+            var text = confirm("Do you want to delete");
+            if(!text) {
+                return false;
+            }
+        }
+    </script>
 </body>
 </html>
